+++ src/loki/reach.cc
@@ -7,7 +7,7 @@ namespace loki {
 
 Reach::Reach() : Dijkstras() {
   // Mock up the Location struct with the important stuff missing
-  auto* path_edge = locations_.Add()->add_path_edges();
+  auto* path_edge = locations_.Add()->mutable_correlation()->add_edges();
   path_edge->set_distance(0);
   path_edge->set_begin_node(false);
   path_edge->set_end_node(false);
@@ -177,9 +177,9 @@ directed_reach Reach::exact(const valhalla::baldr::DirectedEdge* edge,
   auto ll = node->latlng(tile->header()->base_ll());
   locations_.Mutable(0)->mutable_ll()->set_lng(ll.first);
   locations_.Mutable(0)->mutable_ll()->set_lat(ll.second);
-  locations_.Mutable(0)->mutable_path_edges(0)->set_graph_id(edge_id);
-  locations_.Mutable(0)->mutable_path_edges(0)->mutable_ll()->set_lng(ll.first);
-  locations_.Mutable(0)->mutable_path_edges(0)->mutable_ll()->set_lat(ll.second);
+  locations_.Mutable(0)->mutable_correlation()->mutable_edges(0)->set_graph_id(edge_id);
+  locations_.Mutable(0)->mutable_correlation()->mutable_edges(0)->mutable_ll()->set_lng(ll.first);
+  locations_.Mutable(0)->mutable_correlation()->mutable_edges(0)->mutable_ll()->set_lat(ll.second);
 
   // fake up the costing array
   sif::mode_costing_t costings;