+++ src/thor/route_matcher.cc
@@ -24,7 +24,7 @@ using namespace valhalla::thor;
 
 namespace {
 
-using end_edge_t = std::pair<valhalla::Location::PathEdge, float>;
+using end_edge_t = std::pair<valhalla::PathEdge, float>;
 using end_node_t = std::unordered_map<GraphId, end_edge_t>;
 
 // Data type to record edges and transitions that have been followed for each shape
@@ -62,7 +62,7 @@ float length_comparison(const float length, const bool exact_match) {
 // of each edge. This is used to terminate the edge walking method.
 end_node_t GetEndEdges(GraphReader& reader, const valhalla::Location& destination) {
   end_node_t end_nodes;
-  for (const auto& edge : destination.path_edges()) {
+  for (const auto& edge : destination.correlation().edges()) {
     // If destination is at a node - skip any outbound edge
     GraphId graphid(edge.graph_id());
     if (edge.begin_node() || !graphid.Is_Valid()) {
@@ -263,7 +263,7 @@ valhalla::baldr::TimeInfo init_time_info(valhalla::baldr::GraphReader& reader,
   // We support either the epoch timestamp that came with the trace point or
   // a local date time which we convert to epoch by finding the first timezone
   auto time_info = TimeInfo::invalid();
-  for (const auto& e : options.locations(0).path_edges()) {
+  for (const auto& e : options.locations(0).correlation().edges()) {
     GraphId graphid(e.graph_id());
     if (!graphid.Is_Valid() || !reader.GetGraphTile(graphid, tile))
       continue;
@@ -336,7 +336,7 @@ bool RouteMatcher::FormPath(const sif::mode_costing_t& mode_costing,
 
   // Perform the edge walk by starting with one of the candidate edges and walking from it
   // if that walk fails we fall back to another candidate edge until we exhaust the candidates
-  for (const auto& edge : options.locations().begin()->path_edges()) {
+  for (const auto& edge : options.locations().begin()->correlation().edges()) {
     // If origin is at a node - skip any inbound edge
     if (edge.end_node()) {
       continue;
@@ -439,14 +439,14 @@ bool RouteMatcher::FormPath(const sif::mode_costing_t& mode_costing,
           // finding the path but it means that once we do find the path to that node, the edge that
           // was in the value portion of the map entry might be the wrong edge. So here we need to
           // go find the edge candidate that was actually used in the path
-          auto found_edge = destination.path_edges().end();
+          auto found_edge = destination.correlation().edges().end();
           if (n->second.first.end_node()) {
-            found_edge =
-                std::find_if(destination.path_edges().begin(), destination.path_edges().end(),
-                             [&path_infos](const auto& e) -> bool {
-                               return e.graph_id() == path_infos.back().edgeid;
-                             });
-            if (found_edge == destination.path_edges().end()) {
+            found_edge = std::find_if(destination.correlation().edges().begin(),
+                                      destination.correlation().edges().end(),
+                                      [&path_infos](const auto& e) -> bool {
+                                        return e.graph_id() == path_infos.back().edgeid;
+                                      });
+            if (found_edge == destination.correlation().edges().end()) {
               throw std::logic_error("Could not find destination candidate in shape-walked path");
             }
           }
@@ -454,9 +454,10 @@ bool RouteMatcher::FormPath(const sif::mode_costing_t& mode_costing,
           // TODO: when we actually have more than one leg, we have to do when we break legs
           // Store the matching edge candidates in the shapes locations
           const auto& end_edge =
-              found_edge == destination.path_edges().end() ? n->second.first : *found_edge;
-          options.mutable_shape(0)->mutable_path_edges()->Add()->CopyFrom(edge);
-          options.mutable_shape()->rbegin()->mutable_path_edges()->Add()->CopyFrom(end_edge);
+              found_edge == destination.correlation().edges().end() ? n->second.first : *found_edge;
+          options.mutable_shape(0)->mutable_correlation()->mutable_edges()->Add()->CopyFrom(edge);
+          options.mutable_shape()->rbegin()->mutable_correlation()->mutable_edges()->Add()->CopyFrom(
+              end_edge);
 
           // If the end edge is at a node then we are done (no partial time along a destination edge)
           if (end_edge.end_node()) {